refactor: move application page
