+++ apps/admin-ui/src/spa/applications/ValueBox.tsx
