+++ apps/admin-ui/src/spa/applications/queries.ts
