+++ apps/admin-ui/src/spa/recurring-reservations/application-rounds/[id]/review/ApplicationEventsTable.tsx
@@ -10,7 +10,7 @@ import { formatters as getFormatters } from "common";
 import { publicUrl } from "@/common/const";
 import { truncate } from "@/helpers";
 import { applicationDetailsUrl } from "@/common/urls";
-import { getApplicantName } from "@/component/applications/util";
+import { getApplicantName } from "@modules/application";
 import { formatNumber } from "@/common/util";
 import { CustomTable, ExternalTableLink } from "@/component/Table";
 import { ApplicationEventStatusCell } from "./StatusCell";
